version number is removed from output
The version number is interesting to take note of, but it ends up being
changed by the networks on occasions where there has been no actual data
changed.  This ends up muddying the versions stored.  Additionall, the
autoscan is not run often enough to capture all version number transitions
and thus the pattern of use can not be observed.

TODO - extract all the version numbers into a separate file and track
them that way.